requirements.txt
Pillow
pytesseract
PySide6
ttkthemes
opencv-python
